hop by hop header handling moved to the proxy header middleware
