+++ internal/fiber/middleware/proxyheader/hop_by_hop_headers.go
@@ -1,4 +1,4 @@
-package requestcontext
+package proxyheader
 
 import (
 	"net/textproto"
